use CannyFastMath in various places even where it might not be any different
also update a bunch of packages

clean up redundant YamlDotNet references

+++ Content.Server/GameObjects/Components/Chemistry/SolutionComponent.cs
@@ -488,7 +488,7 @@ namespace Content.Server.GameObjects.Components.Chemistry
             //Trigger reaction effects
             foreach (var effect in reaction.Effects)
             {
-                effect.React(Owner, unitReactions.Decimal());
+                effect.React(Owner, unitReactions.Double());
             }
 
             //Play reaction sound client-side
